more inconsistencies in the engine

+++ engine/logging.cpp
@@ -6,12 +6,12 @@ const char *logLevel_0 = "[ImchadaEngine]: ";
 const char *logLevel_1 = "[ImchadaEngine] [ERROR]: ";
 const char *logLevel_2 = "[ImchadaEngine] [WARN]: ";
 
-void imchada_log(std::string log_message, unsigned short log_level) {
+void imchada_log(std::string log_message, int log_level) {
 
   // print_header(), <message>
 }
 
-void imchada_log(char *message, unsigned short log_level) {
+void imchada_log(char *message, int log_level) {
 
   // print_header(), <message>
 }
